Backend/requirements.txt
alembic==1.15.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.54.1
fsspec==2025.5.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.10.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.1.2
orjson==3.10.15
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
streamlit==1.39.0
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.52.4
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
watchdog==5.0.3
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
wordcloud==1.9.3
